File Manager, Resource Manager and Memory Manager
hle/ now answers the Carbon calls the game makes for its files.

- Carbon's File Manager structures keep 68K alignment on Intel, and the
  game's own code proves it: CInfoPBRec.ioNamePtr is at offset 18 and
  ioDrParID at 100. carbon.h packs to 2 bytes and asserts both.
- There is one volume, /. The home and system folders FSFindFolder returns
  live under a stand-in home and root.
- FSRefs and FSSpecs map to paths: classic partial pathnames, node IDs for
  folders and files, and names converted between MacRoman or UTF-16 and
  UTF-8 with '/' and ':' swapped, as Mac OS X does.
- Catalog information, directory iteration in HFS order with AppleDouble
  companions hidden, create, delete and rename, volume information, and
  the parameter-block calls the game uses.
- Forks read and write with pread and pwrite at a per-fork mark. A
  resource fork comes from the file's AppleDouble companion (._name), which
  is where EULA.rsrc keeps its resources.
- The Resource Manager reads resource maps from those forks: GetResource,
  GetIndString and the open-file chain.
- The Memory Manager's pointers and handles, 16-byte aligned like Mac OS X
  malloc.

libmac and hle/ build with a 64-bit off_t, as Darwin's is. That also gives
libmac's Darwin stat structure its real layout.

tests/files_test covers these calls, and with a bundle path it opens the
game's EULA.rsrc resource fork. The game now runs past its file setup and
stops at GetTime.

+++ libmac/mac.c
@@ -469,8 +469,8 @@ int vm_deallocate() {
 
 void *__darwin_mmap(void *addr, size_t length, int prot, int flags,
                     int fd, off_t offset) {
-  LOGF("mmap: addr=%p length=%lu prot=%d flags=%d fd=%d offset=%lu\n",
-       addr, (unsigned long)length, prot, flags, fd, offset);
+  LOGF("mmap: addr=%p length=%lu prot=%d flags=%d fd=%d offset=%lld\n",
+       addr, (unsigned long)length, prot, flags, fd, (long long)offset);
 
   // MAP_ANON is 0x1000 on darwin but 0x20 on linux.
   //
